preparing mbrl tests
